requirements.txt
anyio==4.9.0
attrs==24.2.0
Automat==24.8.1
bcrypt==4.2.0
certifi==2025.1.31
cffi==1.17.1
constantly==23.10.4
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pyseccomp==0.1.2
pyserial==3.5
python-gsmmodem-new==0.13.0
python-telegram-bot==22.0
service-identity==24.1.0
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
Twisted==24.7.0
typing_extensions==4.12.2
zope.interface==7.0.3
